WIP refactor bonsai package
