--- frontend/web/js/bonsai.ts
@@ -1,287 +1,0 @@
-import * as bootstrap from "bootstrap";
-import jQuery from "jquery";
-
-import { initToast, initTooltip, throwSmallToast } from "./utils/notification";
-import { ApiService, AuthService, HttpClient } from "./core/api";
-import { initSamplesTable } from "./utils/table-controller";
-import {
-  deleteSelectedSamples,
-  findAndClusterSimilarSamples,
-  getSimilarSamplesAndCheckRows,
-  initSetSampleQc,
-  removeSamplesFromGroup,
-  updateQcStatus,
-} from "./core/actions/sample-actions";
-import { clusterSamples } from "./core/actions/cluster-actions";
-import { GroupList } from "./components/group-list";
-import { GroupSelector } from "./components/group-selector";
-import { User } from "./core/models/User";
-import { BasketState } from "./core/state/basket-state";
-import { SampleBasketCounter } from "./components/samples-basket-counter";
-import { BasketComponent } from "./components/sample-basket";
-
-import "./components/group-list";
-import "./components/group-selector";
-import "./components/spinner-element";
-import "./utils/choice-select";
-import DataTable from "datatables.net-bs5";
-
-const sampleTableConfig = {
-  select: true,
-  layout: {
-    top1Start: {buttons: ["colvis"]},
-    top2Start: {
-      buttons: ["selectAll", "selectNone", "showSelected", "colvis"],
-    },
-    top2End: {
-      buttons: ["copy", "csv", "excel"],
-    },
-    top3Start: ["searchBuilder"],
-  },
-  lengthMenu: [10, 25, 50, 100, { label: "All", value: -1 }],
-  scrollX: true,
-  pageLength: 50,
-};
-/* Initialize sample basket */
-function initBasket(api: ApiService): BasketState | void {
-  const basketElement = document.querySelector("#basket-content") as HTMLElement;
-  const counterElement = document.querySelector("#basket-counter-container") as HTMLElement;
-  if (!basketElement && !counterElement) {
-    console.error("No DOM element for the basket found!");
-    return;
-  }
-
-  const basketState = new BasketState();
-  // init number of samples in basket counter
-  const basketCounter = new SampleBasketCounter();
-  basketCounter.basketState = basketState;
-  counterElement.appendChild(basketCounter);
-
-  // init component for samples in basket
-  const basketComponent = new BasketComponent(basketState, api.getSamplesDetails.bind(api));
-  basketElement.appendChild(basketComponent);
-
-  // assign functions to DOM objects
-  const clusterBtns = document.querySelectorAll(
-    "#basket-cluster-samples a",
-  ) as NodeListOf<HTMLLinkElement>;
-  clusterBtns.forEach((element) => {
-    element.onclick = () => clusterSamples(element, basketState.getSampleIds(), api);
-  });
-
-  const clearBasketBtn = document.getElementById("clear-basket-btn") as HTMLButtonElement;
-  if (clearBasketBtn) {
-    clearBasketBtn.onclick = () => {
-      basketState.clear();
-    };
-  }
-
-  // setup listeners for rendering sample basket when opening it
-  const offcanvas = document.getElementById("basket-offcanvas");
-  if (offcanvas) {
-    offcanvas.addEventListener("show.bs.offcanvas", () => {
-      basketComponent.render(); // Assuming you have an instance named sampleBasket
-    });
-  }
-  return basketState;
-}
-
-function initApi(bonsaiApiUrl: string, accessToken: string, refreshToken: string) {
-  const auth = new AuthService(bonsaiApiUrl);
-  auth.setTokens(accessToken, refreshToken);
-  const http = new HttpClient(bonsaiApiUrl, auth);
-  const api = new ApiService(http);
-  return api;
-}
-
-/* Initialize interactive elements for the group view. */
-export async function initGroupView(
-  groupId: string | null,
-  bonsaiApiUrl: string,
-  accessToken: string,
-  refreshToken: string,
-): Promise<void> {
-  // setup base functionality
-  const api = initApi(bonsaiApiUrl, accessToken, refreshToken);
-  const basket = initBasket(api);
-  if (!basket) {
-    console.error("Something went wrong when initialize the basket");
-    return;
-  }
-  const headers = document.querySelectorAll<HTMLTableCellElement>("#sample-table thead td");
-  const tableConfig = { ...sampleTableConfig };
-  headers.forEach((cell, idx) => {
-    if (cell.textContent?.trim() === "Date") {
-      tableConfig["order"] = [[idx, "desc"]];
-    }
-  });
-  const table = initSamplesTable("sample-table", tableConfig);
-  // get logged in user
-  const userInfo = await api.getUserInfo();
-  const user = new User(userInfo);
-  initToast();
-  initTooltip();
-
-  // render groups component
-  const groupList = document.createElement("group-list") as GroupList;
-  groupList.getGroupInfo = api.getGroups;
-  groupList.isAdmin = user.isAdmin;
-
-  const groupContainer = document.getElementById("group-container") as HTMLElement;
-  if (groupContainer) groupContainer.appendChild(groupList);
-
-  // attach function to DOM element
-  const addToBasketBtn = document.getElementById("add-to-basket-btn") as HTMLButtonElement;
-  if (addToBasketBtn) addToBasketBtn.onclick = () => basket.addSamples(table.getSelectedRows());
-
-  const deleteSamplesBtn = document.getElementById("remove-samples-btn") as HTMLButtonElement;
-  if (deleteSamplesBtn) deleteSamplesBtn.onclick = () => deleteSelectedSamples(table, api);
-
-  // lookup samples in group if provided a groupId
-  let filterBySamples = null;
-  if (!(groupId === null || groupId === "")) {
-    const group = await api.getGroup(groupId);
-    if ( group.sample_count > 0 ) {
-      try {
-        const controller = new AbortController();
-        const edges = await api.getMembershipByGroups([groupId], controller.signal);
-        // deduplicate ids and default to null if empty
-        const ids = Array.from(new Set(edges.map( e => e.sample_id )));
-        filterBySamples = ids.length === 0 ? null : ids
-      } catch (err) {
-        console.error("Failed to load sample-group memberships", err)
-      }
-    }
-  }
-
-  const selectSimilarSamplesBtn = document.getElementById(
-    "select-similar-samples-btn",
-  ) as HTMLButtonElement;
-  if (selectSimilarSamplesBtn)
-    selectSimilarSamplesBtn.onclick = () =>
-      getSimilarSamplesAndCheckRows(selectSimilarSamplesBtn, table, api, filterBySamples);
-
-  // setup add samples to group component
-  const groupSelectorContainer = document.getElementById(
-    "add-samples-to-group-container",
-  ) as HTMLElement;
-  if (groupSelectorContainer) {
-    const addToGroupSelector = document.createElement("group-selector") as GroupSelector;
-    addToGroupSelector.getGroupInfo = api.getGroups;
-    addToGroupSelector.getSelectedSamples = table.getSelectedRows.bind(table);
-    addToGroupSelector.getGroupMembership = api.getMembershipsBySamples;
-    addToGroupSelector.addToGroup = api.addSamplesToGroup;
-    addToGroupSelector.removeFromGroup = api.removeSamplesFromGroup;
-
-    // set up listners for updates and failurs
-    addToGroupSelector.addEventListener("apply:success", (ev: Event) => {
-      const { groupIds, sampleIds } = (ev as CustomEvent).detail;
-      throwSmallToast(`Added ${sampleIds.length} samples to ${groupIds.length} groups`, "success");
-    });
-    addToGroupSelector.addEventListener("apply:error", (ev: Event) => {
-      const { error } = (ev as CustomEvent).detail;
-      console.error("Apply error:", error);
-      throwSmallToast(`An error occured: ${ev.detail}`, "error");
-    });
-
-    addToGroupSelector.addEventListener("apply:skipped", (ev: Event) => {
-      const { reason } = (ev as CustomEvent).detail;
-      const msg =
-        reason === "empty"
-          ? "No groups or samples selected"
-          : reason === "no-samples"
-            ? "No samples selected"
-            : reason === "no-groups"
-              ? "No groups selected"
-              : "Nothing to do";
-      throwSmallToast(msg, "warning");
-    });
-
-    groupSelectorContainer.appendChild(addToGroupSelector);
-    // preselect groups when selecting samples in sample table
-    table &&
-      table.getTable().on("select deselect", (e, dt, type, indexes) => {
-        const selected: string[] = dt.rows(".selected").ids().toArray();
-        addToGroupSelector.preselectGroupsForSamples(selected);
-      });
-  }
-
-  // setup qc classification
-  const qcStatusForm = document.getElementById("qc-form-control") as HTMLButtonElement;
-  if (qcStatusForm)
-    initSetSampleQc(
-      table.getSelectedRows.bind(table),
-      api.setSampleQc.bind(api),
-      () => console.log("table needs to be redrawn"),
-      qcStatusForm,
-    );
-  // FIXME updating individual cells did not work for some reason
-  // the entire table might need to be redrawn.
-
-  const removeFromGroupBtn = document.getElementById("remove-from-group-btn") as HTMLButtonElement;
-  if (removeFromGroupBtn)
-    removeFromGroupBtn.onclick = () => {
-      const groupId: string = removeFromGroupBtn.getAttribute("data-bi-group-id");
-      removeSamplesFromGroup(groupId, table, api);
-    };
-}
-
-export async function initVariantsTable(tableId: string, search: boolean = true): Promise<Any> {
-  if (document.getElementById(tableId) === null) {
-    console.warn(`No table with id: ${tableId} found, cant create datatable`);
-    return;
-  }
-  console.log(tableId);
-  const tbl = new DataTable(tableId, {
-    paging: false,
-    select: false,
-    searching: search,
-  });
-  return tbl;
-}
-
-/* Initialize interactive elements for the sample view. */
-export async function initSampleView(
-  bonsaiApiUrl: string,
-  accessToken: string,
-  refreshToken: string,
-  sampleId: string,
-  groupId: string | null,
-): Promise<string> {
-  // setup base functionality
-  const api = initApi(bonsaiApiUrl, accessToken, refreshToken);
-  initToast();
-  initTooltip();
-
-  const qcStatusForm = document.getElementById("qc-classification-form") as HTMLButtonElement;
-  if (qcStatusForm) {
-    initSetSampleQc(() => [sampleId], api.setSampleQc.bind(api), updateQcStatus, qcStatusForm);
-  }
-  // lookup samples in group if provided a groupId
-  let narrow_search_to = null;
-  if (groupId !== null) {
-    const group = await api.getGroup(groupId);
-    narrow_search_to = group.included_samples.length > 0 ? group.included_samples : null;
-  }
-  const newick = await findAndClusterSimilarSamples(sampleId, narrow_search_to, api);
-  return newick;
-}
-
-declare global {
-  interface Window {
-    throwSmallToast: (message: string) => void;
-    jQuery: typeof jQuery;
-    $: typeof jQuery;
-    bootstrap: typeof bootstrap;
-  }
-  interface HTMLElementTagNameMap {
-    "groups-list": GroupList;
-    "choice-select": ChoiceSelect;
-    "group-selector": GroupSelector;
-  }
-}
-
-window.throwSmallToast = throwSmallToast;
-window.jQuery = jQuery;
-window.$ = jQuery;
-window.bootstrap = bootstrap;
