+++ frontend/web/js/views/groups-view/index.ts
@@ -1,3 +1,1 @@
-import { initGroupView } from "./init";
-
-initGroupView();
+export { initGroupView } from "./init";
