+++ frontend/web/js/views/sample-view/index.ts
@@ -1,0 +1,1 @@
+export { initSampleView } from "./init";
