+++ frontend/web/js/views/variants-table/index.ts
@@ -1,0 +1,1 @@
+export { initVariantsTable } from "./init";
